sheesh

+++ folders/project1_material/project_one/aseel/child.c
@@ -5,9 +5,11 @@
 #include <signal.h>
 #include <string.h>
 void signal_catcher(int);
+void round_reset_func(int);
 int distance = 50 ;
 int teamFlag =0;
 int speed = 0;
+int reset_round = 0 ;
 int main(int argc, char *argv[])
 {
     speed = 5;
@@ -15,17 +17,19 @@ int main(int argc, char *argv[])
     sleep(1);
     kill(getppid(),10);
     if(strcmp("shit", *argv) == 0){
-        printf("i am in team one %d\n", getpid());
         teamFlag = 1 ;
     }
     else{
-        printf("i am in team two\n");
         teamFlag = 2 ;
     }
     // printf("Invoked CHILD: My process ID is %d \n", getppid());
 
     while(1){
     if(sigset(3, signal_catcher) == -1){
+        perror("Sigset can not set SIGINT");
+        exit(SIGINT);
+    }
+    if(sigset(15,round_reset_func) == -1){
         perror("Sigset can not set SIGINT");
         exit(SIGINT);
     }
@@ -43,19 +47,29 @@ int main(int argc, char *argv[])
 
 void signal_catcher(int sig){
     int speed = 3+getpid()%5;
-    
-    while(distance > 0){
-        printf("Child %d is running at speed %d \n", getpid(), speed);
+    reset_round = 0;
+    distance  = 50;
+    while(distance > 0 && reset_round == 0){
+        printf("Player from team %d with pid %d is running at speed %d left distance is %d \n", teamFlag, getpid(),speed,distance);
         distance -= speed;
         sleep(1);
     }
-    if(teamFlag == 1){
+    if(teamFlag == 1 && reset_round == 0) {
         kill(getppid(), 12);
+        printf("Player from team %d with pid %d finsihed his round\n",teamFlag,getpid());
     }
-    else{
+    else if (teamFlag == 2 && reset_round == 0){
         kill(getppid(), 13);
+        printf("Player from team %d with pid %d finsihed his round\n",teamFlag,getpid());
+    }
+    else if ( reset_round == 1){
+        printf("the other team has won the game waiting for the next round \n");
     }
 
-    printf("\nholyshit\n");
 }
 //  generate random number between 5 and 8
+
+
+void round_reset_func(int sig){
+    reset_round = 1 ;
+}
